building ui: stuff data into main_ui

+++ lib/main_ui/short_forecast/model/percent.dart
@@ -7,7 +7,7 @@ class Percent {
   int dbz;
   String desc;
   int icon;
-  int percent;
+  double percent;
 
   Percent({
     this.dbz,
